Move and consolidate repr tests
